added tests for circleci + dotenv to use a .env file, for apikey and db connection
