+++ start.js
@@ -2,6 +2,7 @@
 require("@babel/register")({
     presets: ["@babel/preset-env"]
 });
-  
+require('dotenv').config()
+
 // Import the rest of our application.
 module.exports = require('./server.js')
